#68 Make FindPwTopBar that is separated from FindPw  and Export FindPwChildProps from it so Use it instead of FindPwContentsProps and Delete it
